ModelRetrySettingsTab and PriceDataManagementTab
